Add type attribute, mobile IE11 stabilisation
